+++ apis/gear/src/utils/address.ts
@@ -1,6 +1,7 @@
-import { HexString } from '@polkadot/util/types';
 import { Keyring } from '@polkadot/api';
 import { u8aToHex } from '@polkadot/util';
+
+import { HexString } from '../types';
 
 export function decodeAddress(publicKey: string): HexString {
   return u8aToHex(new Keyring().decodeAddress(publicKey));