+++ idea/vara-eth/frontend/src/features/codes/lib/queries.ts
@@ -1,20 +1,17 @@
 import { useQuery } from '@tanstack/react-query';
+import { HexString } from '@vara-eth/api';
 
 import { EXPLORER_URL } from '@/shared/config';
 import { PaginatedResponse } from '@/shared/types';
 import { fetchWithGuard } from '@/shared/utils';
+
+import { Code, getCode } from './requests';
 
 export const CODE_STATUS = {
   VALIDATION_REQUESTED: 'ValidationRequested',
   VALIDATION_FAILED: 'ValidationFailed',
   VALIDATED: 'Validated',
 } as const;
-
-export type Code = {
-  id: string;
-  status: string;
-  createdAt: string;
-};
 
 type CodesResponse = PaginatedResponse<Code>;
 
@@ -35,10 +32,10 @@ export const useGetAllCodesQuery = (page: number, pageSize: number) => {
   });
 };
 
-export const useGetCodeByIdQuery = (id: string) => {
+export const useGetCodeByIdQuery = (id: HexString | undefined) => {
   return useQuery({
     queryKey: ['codeById', id],
-    queryFn: () => fetchWithGuard<Code>({ url: `${EXPLORER_URL}/codes/${id}` }),
-    enabled: !!id,
+    queryFn: () => getCode(id!),
+    enabled: Boolean(id),
   });
 };