+++ idea/vara-eth/frontend/src/features/programs/lib/hooks/index.ts
@@ -4,3 +4,4 @@ export { useInitProgram } from './use-init-program';
 export { useReadContractState } from './use-read-contract-state';
 export { useSails } from './use-sails';
 export { useSendProgramMessage } from './use-send-program-message';
+export { useSendInjectedTransaction } from './use-send-injected-transaction';