+++ idea/vara-eth/frontend/src/features/programs/lib/index.ts
@@ -5,6 +5,7 @@ export {
   useReadContractState,
   useSails,
   useSendProgramMessage,
+  useSendInjectedTransaction,
 } from './hooks';
 export { useGetAllProgramsQuery, useGetProgramByIdQuery } from './queries';
 export { type PayloadValue } from './types';