+++ idea/vara-eth/frontend/src/features/programs/lib/queries.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { type Code } from '@/features/codes/lib/queries';
+import { type Code } from '@/features/codes/lib/requests';
 import { EXPLORER_URL } from '@/shared/config';
 import { PaginatedResponse } from '@/shared/types';
 import { fetchWithGuard } from '@/shared/utils';