+++ idea/vara-eth/frontend/src/features/programs/ui/message-form/message-form.tsx
@@ -6,7 +6,7 @@ import { Sails } from 'sails-js';
 
 import { Button, Input, ExpandableItem } from '@/components';
 
-import { PayloadValue, useSendProgramMessage } from '../../lib';
+import { PayloadValue, useSendInjectedTransaction, useSendProgramMessage } from '../../lib';
 import { ISailsFuncArg } from '../../lib/types';
 import { getDefaultPayloadValue, getPayloadSchema } from '../../lib/utils';
 
@@ -19,14 +19,20 @@ type Props = {
   isQuery: boolean;
   sails: Sails;
   args: ISailsFuncArg[];
+  idl: string;
+  isOffchain: boolean;
 };
 
 type Values = {
   [k: string]: PayloadValue;
 };
 
-const MessageForm = ({ programId, isQuery, sails, serviceName, messageName, args }: Props) => {
-  const { sendMessage, isPending } = useSendProgramMessage(programId);
+const MessageForm = ({ programId, isQuery, sails, serviceName, messageName, args, idl, isOffchain }: Props) => {
+  const { mutate: sendInjectedTransaction, isPending: isPendingInjectedTransaction } = useSendInjectedTransaction(
+    programId,
+    idl,
+  );
+  const { mutate: sendMessage, isPending: isPendingMessage } = useSendProgramMessage(programId, idl);
 
   const defaultValues = useMemo(() => getDefaultPayloadValue(sails, args), [sails, args]);
 
@@ -39,7 +45,8 @@ const MessageForm = ({ programId, isQuery, sails, serviceName, messageName, args
   };
 
   const handleSubmitForm = form.handleSubmit((formValues) => {
-    sendMessage({ serviceName, messageName, args: formValues, isQuery }, { onSuccess: resetForm });
+    const send = isOffchain ? sendInjectedTransaction : sendMessage;
+    send({ serviceName, messageName, args: formValues, isQuery }, { onSuccess: resetForm });
   });
 
   const formName = `${serviceName}-${messageName}-form`;
@@ -57,7 +64,7 @@ const MessageForm = ({ programId, isQuery, sails, serviceName, messageName, args
               variant="default"
               form={formName}
               size="xs"
-              isLoading={isPending}
+              isLoading={isPendingInjectedTransaction || isPendingMessage}
               className={styles.button}>
               {isQuery ? 'Read' : 'Write'}
             </Button>