+++ idea/vara-eth/frontend/src/features/programs/ui/service-list/service-list.tsx
@@ -1,21 +1,25 @@
 import { HexString } from '@vara-eth/api';
+import { useState } from 'react';
 
-import { Badge, ExpandableItem } from '@/components';
+import { Badge, ExpandableItem, Tabs } from '@/components';
 
-import counterIdl from '../../../../../../../../apis/vara-eth/programs/counter-idl/counter.idl?raw';
 import { useReadContractState, useSails } from '../../lib';
 import { InitForm } from '../init-form';
 import { MessageForm } from '../message-form';
 
 import styles from './service-list.module.scss';
 
+const tabs = ['Call offchain', 'Call onchain'];
+
 type Props = {
   programId: HexString;
+  idl: string;
 };
 
-const ServiceList = ({ programId }: Props) => {
+const ServiceList = ({ programId, idl }: Props) => {
   const { data: programState, refetch, isPending } = useReadContractState(programId);
-  const { data: sails } = useSails(counterIdl);
+  const { data: sails } = useSails(idl);
+  const [tabIndex, setTabIndex] = useState(0);
 
   const isInitialized = programState && 'Active' in programState.program && programState.program.Active.initialized;
 
@@ -39,6 +43,13 @@ const ServiceList = ({ programId }: Props) => {
     <div>
       {isInitialized ? (
         <>
+          <Tabs
+            tabs={tabs}
+            tabIndex={tabIndex}
+            onTabIndexChange={(index) => setTabIndex(index)}
+            className={styles.tabs}
+          />
+
           {services.map(([serviceName, service]) => {
             const functions = Object.entries(service.functions);
             const queries = Object.entries(service.queries);
@@ -66,6 +77,8 @@ const ServiceList = ({ programId }: Props) => {
                       sails={sails}
                       args={args}
                       isQuery={isQuery}
+                      idl={idl}
+                      isOffchain={tabIndex === 0}
                     />
                   );
                 })}
@@ -90,6 +103,7 @@ const ServiceList = ({ programId }: Props) => {
               ctorName={ctorName}
               args={args}
               onInit={refetch}
+              idl={idl}
             />
           ))}
         </ExpandableItem>