+++ idea/vara-eth/frontend/src/shared/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useChangeEffect } from './use-change-effect';
+export { useIdlStorage } from './use-idl-storage';